requirements.txt
ccxt==4.1.97
websockets==12.0
pyyaml==6.0.1
pandas==2.2.2
plotly==5.20.0
streamlit==1.32.0
colorama==0.4.6
